backend/requirements.txt
Django==4.2
# asgiref
# distlib
# filelock
# platformdirs
# sqlparse
# tzdata
# virtualenv
pillow
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
celery
